vmxnet3: remove the special case rule for vmxnet3
There was a hotplug rule for vmxnet3 that forced an unbind/rebind
of the vmxnet3 interfaces as sometimes they came up with uio as the
driver instead of vmxnet3. Depending on the timing of this (with
respect to dataplane start) it could cause it to look like a hotplug
reattach, and mess up the state for the interface. Remove this
workaround as it is no longer needed.

VRVDR-49420